planettool: detect overflow when specifying ridiculously large output sizes. Unified version numbering between command-line and Mac GUI versions. Bumped version to 0.4.3.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk/tools/planettool@3253 127b21dd-08f5-0310-b4b7-95ae10353056

+++ RenderToCube.h
@@ -28,6 +28,6 @@
 #include "SphericalPixelSource.h"
 
 
-FloatPixMapRef RenderToCube(size_t size, RenderFlags flags, SphericalPixelSourceFunction source, void *sourceContext, ProgressCallbackFunction progress, ErrorCallbackFunction error, void *cbContext);
+FloatPixMapRef RenderToCube(uintmax_t size, RenderFlags flags, SphericalPixelSourceFunction source, void *sourceContext, ProgressCallbackFunction progress, ErrorCallbackFunction error, void *cbContext);
 
-FloatPixMapRef RenderToCubeCross(size_t size, RenderFlags flags, SphericalPixelSourceFunction source, void *sourceContext, ProgressCallbackFunction progress, ErrorCallbackFunction error, void *cbContext);
+FloatPixMapRef RenderToCubeCross(uintmax_t size, RenderFlags flags, SphericalPixelSourceFunction source, void *sourceContext, ProgressCallbackFunction progress, ErrorCallbackFunction error, void *cbContext);
